+++ integration-tests/src/test/resources/environment-configuration.yml
@@ -104,3 +104,5 @@ instanceEnv:
     ZWE_configs_apiml_security_auth_jwt_customAuthHeader: customJwtHeader
     ZWE_configs_apiml_security_auth_passticket_customUserHeader: customUserHeader
     ZWE_configs_apiml_security_auth_passticket_customAuthHeader: customPassticketHeader
+    ZWE_configs_apiml_service_customMetadata_apiml_lb_type: # set the value to "authentication" if you want to test the sticky session load balancing
+
